rework schema code
